Fix the ">" not showing up
The ">" symbol used to indicate the currently selected results, but it stopped working. I could get it to show up if I set the position to "left:0px" instead of "left:-15px", but then the ">" symbol appeared on top of other text. I went a totally different route and let focus() do the job, ensuring the page scrolls down to the selected element, and hopefully insulates us from future changes since we're not styling anything now.

Also, explicitly filter out Ads (which weren't a problem before) and the "People also ask" section, which means the keyboard shortcut jumps from one link to another, without having to press the Down Arrow multiple times to skip that section.