fix 'featured snippets' links (#1)
Sometimes google results have a special link at the top of the page called a 'featured snippet', that includes text from the page google thinks answers your question, plus a link to the page. That's often the most relevant link but the code was skipping it. Now we don't, and we still skip the 'People also ask' section (which has hidden links, so we don't want to put the focus on them).